requirements.txt
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httplib2==0.22.0